third_party/WebKit/LayoutTests/fast/filesystem/input-access-entries.html

<!DOCTYPE html>
<html>
<body>
<input name="input" id="input"></input>
<script src="../../resources/js-test.js"></script>
<script>
description('Tests accessing webkitEntries does not crash when input type is not file');

var input = document.getElementById('input');
shouldEvaluateTo('input.webkitEntries.length', 0);

</script>
</body>
</html>
